fix: move timer to its own component
